% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find.loop.nodes}
\alias{find.loop.nodes}
\title{Find nodes that form loops.}
\usage{
find.loop.nodes(adjacency, IGRAPH = FALSE)
}
\arguments{
\item{adjacency}{A square adjacency matrix.}

\item{IGRAPH}{if \code{IGRAPH} is true adjacency is already a Igraph object not a matrix.}
}
\value{
The row/col numbers for nodes that form loops in the graph.
}
\description{
Find nodes that form loops.
}
\examples{
data( models )
find.loop.nodes( sim$matrix )
}
\author{
David O'Sullivan
}

