% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.DAG.collaspe}
\alias{generate.DAG.collaspe}
\title{Find the DAG of a graph.}
\usage{
generate.DAG.collaspe(adjacency)
}
\arguments{
\item{adjacency}{A square adjacency matrix.}
}
\value{
List of three igraph objects (\code{orginal_graph},\code{SCC_graph} and \code{DAG}). First is the
orginal graph, then the nodes that form loops (strongly connected componet) and the DAG for the graph.
}
\description{
Take the adjacency matrix of a graph, identify nodes that form cycles and construct DAG from this system.
}
\examples{
data( models )
generate.DAG.collaspe( sim$matrix )
}
\author{
David O'Sullivan
}

