% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.DAG.collaspe_oldversion}
\alias{generate.DAG.collaspe_oldversion}
\title{Find the DAG of a graph}
\usage{
generate.DAG.collaspe_oldversion(adjacency)
}
\arguments{
\item{adjacency}{The adjacency matrix for the graph}
}
\value{
List of three igraph objects (\code{orginal_graph},\code{SCC_graph} and \code{DAG}). First is the
orginal graph, then the nodes that form loops (strongly connected componet) and the DAG for the graph.
}
\description{
Take the adjacency matrix of a graph, identify nodes that form cycles and construct DAG from this system.
}
\author{
David O'Sullivan
}

