% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.DAG}
\alias{is.DAG}
\title{Is graph a DAG.}
\usage{
is.DAG(adjacency, IGRAPH = FALSE)
}
\arguments{
\item{adjacency}{A square adjacency matrix.}

\item{IGRAPH}{if \code{IGRAPH} is true adjacency is already a Igraph object not a matrix.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks graph to see if it is a DAG or not.
}
\examples{
data( models )
is.DAG( sim$matrix )
g = graph.adjacency( adjmatrix = sim$matrix, mode = "directed" )
is.DAG( g, IGRAPH = TRUE )
}
\author{
David O'Sullivan
}

