% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.sfc}
\alias{is.sfc}
\title{Check if \code{x} is an sfc object.}
\usage{
is.sfc(x)
}
\arguments{
\item{x}{an object to be tested.}
}
\value{
TRUE if x is an object of the class sfc and FALSE otherwise.
}
\description{
Check if \code{x} is an sfc object.
}
\examples{
data( models )
is.sfc( sim )
}
\author{
Antoine Godin
}

