% Generated by roxygen2 (4.0.2): do not edit by hand
\name{min_subset_to_remove_nodes}
\alias{min_subset_to_remove_nodes}
\title{Find the minimumber of nodes to remove from graph to create a DAG.}
\usage{
min_subset_to_remove_nodes(adjacency, root_node = 1, PLOT = FALSE)
}
\arguments{
\item{adjacency}{A square adjacency matrix.}

\item{root_node}{The node that forms the root of the plot.}

\item{PLOT}{If plot is \code{TRUE}, create a plot of the graph with nodes that are to be removed
highlighted in light pink and nodes outside of cycles higlighted in green.}
}
\description{
Find the minimumber of nodes to remove from graph to create a DAG.
}
\examples{
data( models )
min_subset_to_remove_nodes( sim$matrix )
}
\author{
David O'Sullivan
}

