% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.dag}
\alias{plot.dag}
\title{Plot sfc objects as an igraph hierarchy structure}
\usage{
plot.dag(model, ...)

plot.dag(model, ...)
}
\arguments{
\item{model}{an sfc object.}

\item{...}{plot options for graph objects}
}
\description{
Plot sfc objects as an igraph hierarchy structure

}
\examples{
data( models )
plot_graph_hierarchy( model=sim )
}
\author{
David O'Sullivan
}

