% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_graph_hierarchy}
\alias{plot_graph_hierarchy}
\title{Plot igraph objects hierarchy structure}
\usage{
plot_graph_hierarchy(graph, ...)

plot_graph_hierarchy(graph, ...)
}
\arguments{
\item{graph}{an igraph object.}

\item{...}{plot options for graph objects}

\item{graph}{an igraph object.}

\item{...}{plot options for graph objects}
}
\description{
Plot igraph objects hierarchy structure

Plot igraph objects hierarchy structure
}
\examples{
data( models )
g = generate.DAG.collaspe( adjacency = sim$matrix )$orginal_graph
plot_graph_hierarchy( graph = g )
data( models )
g = generate.DAG.collaspe( adjacency = sim$matrix )$orginal_graph
plot_graph_hierarchy( graph = g )
}
\author{
David O'Sullivan

David O'Sullivan
}

