% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.summary.sfc}
\alias{print.summary.sfc}
\title{Print a summary of the sfc model.}
\usage{
\method{print}{summary.sfc}(x, ...)
}
\arguments{
\item{x}{an sfc object.}

\item{...}{further arguments that can be passed to function.}
}
\value{
print values to screen.
}
\description{
print a condensed summary of sfc object.
}
\author{
Antoine Godin
}

