% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.addEnd}
\alias{sfc.addEnd}
\title{Add endogenous values to sfc.}
\usage{
sfc.addEnd(model = stop("Need a model"),
  var = stop("Need a variable name!"), init = NA, lag = 0, desc = NA)
}
\arguments{
\item{model}{an sfc object.}

\item{var}{the endogenous variable to be added.}

\item{init}{the initial condtions for the endogenous variable.}

\item{lag}{which lag is to be used for updating this variable.}

\item{desc}{descriptions of endogenous variable.}
}
\value{
an sfc object.
}
\description{
Add a single endogenous variable to the sfc object.
}
\author{
Antoine Godin
}

