% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.addEnds}
\alias{sfc.addEnds}
\title{A a list of endogenous variables.}
\usage{
sfc.addEnds(model = stop("Need a model"),
  list = stop("Need a list f variables"))
}
\arguments{
\item{model}{an sfc object.}

\item{list}{a list containing the endogenous variables, intial conditions,
lags and description of the endogenous variable.}
}
\value{
an sfc object.
}
\description{
Add a list of endogenous variables to the sfc object.
}
\author{
Antoine Godin
}

