% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.addEqu}
\alias{sfc.addEqu}
\title{Add equation.}
\usage{
sfc.addEqu(model = stop("Need a model"),
  var = stop("Need a variable name!"), eq = stop("Need an equation!"),
  desc = "")
}
\arguments{
\item{model}{an sfc object.}

\item{var}{the varaibles in the eqauation.}

\item{eq}{the equations to be added to the sfc object.}

\item{desc}{a descriptions of the equation.}
}
\value{
an sfc object.
}
\description{
Add a single equation to a sfc object.
}
\author{
Antoine Godin
}

