% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.addEqus}
\alias{sfc.addEqus}
\title{Adding Equations.}
\usage{
sfc.addEqus(model = stop("Need a model"),
  list = stop("Need a list f variables"))
}
\arguments{
\item{model}{an sfc object.}

\item{list}{a list containing the equations, varaibles and descriptions to be added to the
sfc object.}
}
\value{
an sfc object.
}
\description{
Add a series of equations to an sfc object.
}
\author{
Antoine Godin
}

