% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.addScenario}
\alias{sfc.addScenario}
\title{Add a scenario.}
\usage{
sfc.addScenario(model = stop("Need a model"), vars = stop("Need variables"),
  values = stop("Need values"), inits = stop("Need starting values"),
  ends = stop("Need ending values"), starts = NA)
}
\arguments{
\item{model}{an sfc object.}

\item{vars}{variable that is to be changed.}

\item{values}{the new values for the variable.}

\item{inits}{when the scenario should start.}

\item{ends}{when should the scenario end.}

\item{starts}{the inital conditions for the scenario.}
}
\value{
an sfc object.
}
\description{
Add a scenario (such as a shock to the model) to be added to the sfc object.
}
\author{
Antoine Godin
}

