% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.addVar}
\alias{sfc.addVar}
\title{Add variable.}
\usage{
sfc.addVar(model = stop("Need a model"),
  var = stop("Need a variable name!"), init = "NA", desc = "")
}
\arguments{
\item{model}{an sfc object.}

\item{var}{the variable to be added to the sfc object.}

\item{init}{the inital conditions for the variable.}

\item{desc}{a description of the variable.}
}
\value{
an sfc object.
}
\description{
Add a single variable to a sfc object.
}
\author{
Antoine Godin
}

