% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.addVars}
\alias{sfc.addVars}
\title{Add variables.}
\usage{
sfc.addVars(model = stop("Need a model"),
  list = stop("Need a list f variables"))
}
\arguments{
\item{model}{an sfc object.}

\item{list}{a list containing the variables to be added, initial conditions and descriptions.}
}
\value{
an sfc object.
}
\description{
Add a series of variables to a sfc object.
}
\author{
Antoine Godin
}

