% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.check}
\alias{sfc.check}
\title{Check format of sfc object.}
\usage{
sfc.check(model = stop("Need a model"), fill = F)
}
\arguments{
\item{model}{an sfc object.}

\item{fill}{if some values are missing would you like to be prompted to fill them in?}
}
\description{
Run a sanity check for errors in a sfc object.
}
\author{
Antoine Godin
}

