% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.create}
\alias{sfc.create}
\title{create an sfc object.}
\usage{
sfc.create(modelName = stop("Need a model name"))
}
\arguments{
\item{modelName}{Optional, default="SFCmodel". Specifies the name of the
model.}
}
\value{
an sfc object.
}
\description{
Create an emtpy sfc object.
}
\author{
Antoine Godin
}

