% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.editEnd}
\alias{sfc.editEnd}
\title{Edit endogenous variable.}
\usage{
sfc.editEnd(model = stop("Need a model"), ind = NA, var = NA, init = NA,
  lag = NA, desc = NA)
}
\arguments{
\item{model}{an sfc object.}

\item{ind}{index of endogenous variable to be changed.}

\item{var}{the endogenous variable to be added.}

\item{init}{the initial condtions for the endogenous variable.}

\item{lag}{which lag is to be used for updating this variable.}

\item{desc}{descriptions of endogenous variable.}
}
\value{
sfc object.
}
\description{
Edit an endogenous variable of a sfc object.
}
\author{
Antoine Godin
}

