% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.editEnds}
\alias{sfc.editEnds}
\title{Edit endogenous Variables.}
\usage{
sfc.editEnds(model = stop("Need a model"),
  list = stop("Need a list f variables"))
}
\arguments{
\item{model}{an sfc object.}

\item{list}{a list containing lists of the indexs of endogenous variables to be changes,
the variables to be changed, the initial conditions, the lags and the descriptions of the
endogenous variables to be edited.}
}
\value{
an sfc model.
}
\description{
Edit a list of endogenous variables in an sfc object.
}
\author{
Antoine Godin
}

