% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.editEqu}
\alias{sfc.editEqu}
\title{Edit equation.}
\usage{
sfc.editEqu(model = stop("Need a model"), ind = NA, var = NA, eq = NA,
  desc = NA)
}
\arguments{
\item{model}{an sfc object.}

\item{ind}{the index of the equation to be edited.}

\item{var}{the varaibles in the eqauation.}

\item{eq}{the equations to be added to the sfc object.}

\item{desc}{a descriptions of the equation.}
}
\value{
an sfc object.
}
\description{
Edit an equation in an sfc object.
}
\author{
Antoine Godin
}

