% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.editEqus}
\alias{sfc.editEqus}
\title{Edit equations.}
\usage{
sfc.editEqus(model = stop("Need a model"),
  list = stop("Need a list f variables"))
}
\arguments{
\item{model}{an sfc object.}

\item{list}{a list containing lists of the indexs of equations to be changes,
the equations to be changed, the initial conditions, the lags and the descriptions of the
equations to be edited.}
}
\value{
an sfc object.
}
\description{
Edit a series of equations in an sfc object.
}
\author{
Antoine Godin
}

