% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.editVar}
\alias{sfc.editVar}
\title{Edit variable.}
\usage{
sfc.editVar(model = stop("Need a model"), ind = NA, var = NA, init = NA,
  desc = NA)
}
\arguments{
\item{model}{an sfc object.}

\item{ind}{index of the variable to be edited.}

\item{var}{the variable to be added to the sfc object.}

\item{init}{the inital conditions for the variable.}

\item{desc}{a description of the variable.}
}
\value{
an sfc object.
}
\description{
edit the variable of a sfc object .
}
\author{
Antoine Godin
}

