% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.editVars}
\alias{sfc.editVars}
\title{Editing Variables.}
\usage{
sfc.editVars(model = stop("Need a model"),
  list = stop("Need a list f variables"))
}
\arguments{
\item{model}{an sfc model.}

\item{list}{a list containing lists of the indexs of variables to be changes,
the variables to be changed, the initial conditions, the lags and the descriptions of the
variables to be edited.}
}
\value{
an sfc object.
}
\description{
Editin variables in an sfc object.
}
\author{
Antoine Godin
}

