% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.eval}
\alias{sfc.eval}
\title{Evaluate an sfc object.}
\usage{
sfc.eval(object = stop("Need a model to evaluate"), check = F)
}
\arguments{
\item{object}{sfc object}

\item{check}{would you like to run a test for missing information?}
}
\value{
a list of strings consistent with what is required to get the GaussSeidel
algorithm to work.
}
\description{
Reformats an sfc object into a format that can be used by sfc.GaussSeidel.
}
\author{
Antoine Godin
}

