% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.eviews}
\alias{sfc.eviews}
\title{Create an sfc object from an eviews file.}
\usage{
sfc.eviews(fileName, modelName = "SFCmodel")
}
\arguments{
\item{fileName}{The path to the filename containing the equations. See
details hereunder for a description of the format of the text file}

\item{modelName}{Optional, default="SFCmodel". Specifies the name of the
model.}
}
\value{
An sfc object
}
\description{
Given an eviews file, the file will be parsed and a sfc object will be created.
}
\author{
Antoine Godin
}

