% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.GaussSeidel}
\alias{sfc.GaussSeidel}
\title{Run Gauss-Seidel algorithm.}
\usage{
sfc.GaussSeidel(eqs, tol, maxIter, initialValues, prev)
}
\arguments{
\item{eqs}{the equations of the system.}

\item{tol}{the max change between itterations required for convergence.}

\item{maxIter}{the max number of itterations.}

\item{initialValues}{the initial values of the Gauss Seidel algorithm.}

\item{prev}{lag values to be used.}
}
\value{
returns the solution of the Gauss Seidel algorithm for the system
}
\description{
Run the Gauss Seidel algorithm to solve the system of equations.
}
\author{
Antoine Godin
}

