% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.getIndex}
\alias{sfc.getIndex}
\title{Find index.}
\usage{
sfc.getIndex(model = stop("Need a model"), var = NA, eq = NA, end = NA)
}
\arguments{
\item{model}{an sfc object.}

\item{var}{the variable that you want the index for, if applicable.}

\item{eq}{the equation that you want the index for, if applicable.}

\item{end}{the endogenous variable that you want the index for, if applicable.}
}
\value{
the required index.
}
\description{
Find the indexs for sfc object inputs.
}
\author{
Antoine Godin
}

