% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.model}
\alias{sfc.model}
\title{Generating SFC models}
\usage{
sfc.model(fileName, dataFile = NA, modelName = "SFCmodel", fill = F)
}
\arguments{
\item{fileName}{The path to the filename containing the equations. See
details hereunder for a description of the format of the text file}

\item{dataFile}{Optional. The path to the file containing the values for the
variables (endogenous) and (exogenous) of the model. See details hereunder
for a description of the format of the data file}

\item{modelName}{Optional, default="SFCmodel". Specifies the name of the
model.}

\item{fill}{Optional, default FALSE. Wether the check of the model offers to
fill the lacks found in the model. See sfc.check for more details.}
}
\value{
An object of the class "sfc" representing a model
}
\description{
Generate an SFC model from a text file and an eventual database. The
function generates the model and checks it.
}
\details{
The left hand side (LHS) of an equation should contain only one variable
which is considered as endogenous. If the right hand side (RHS) part of the
equation contains only a number, this number is set as the initial value
(resp. fixed value) for endogenous (resp. exogenous) variable contained in
the LHS part of the equation. Each equation might preceded by a description
of the equation. The timeline of the model is specified with the reserved
keyword 'timeline' and two dates separated by spaces.  #Disposable income yd
= w*n - t timeline 1945 2010 The format of the data file (respecting csv
format, 'white space' separator and "." decimal) is as follows: Names of
each variable should be in the first line, years in the first column.
}
\note{
%% ~~further notes~~
}
\author{
Antoine Godin

Antoine Godin
}
\references{
%% here is where the references go
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\keyword{~kwd1}
\keyword{~kwd2}

