% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.print}
\alias{sfc.print}
\title{Print a sfc model.}
\usage{
sfc.print(object)

sfc.print(object)
}
\arguments{
\item{object}{an sfc object.}

\item{object}{an sfc object.}
}
\description{
print sfc model to screen.

print sfc model to screen.
}
\author{
Antoine Godin
}

