% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.rmEnd}
\alias{sfc.rmEnd}
\title{Remove endogenous variable.}
\usage{
sfc.rmEnd(model = stop("Need a model"), ind = stop("Need an index!"))
}
\arguments{
\item{model}{an sfc object.}

\item{ind}{index of endogenous variable to be changed.}
}
\value{
an sfc model.
}
\description{
Remove an endogenous variable from an sfc model.
}
\author{
Antoine Godin
}

