% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.rmEqu}
\alias{sfc.rmEqu}
\title{Remove equation.}
\usage{
sfc.rmEqu(model = stop("Need a model"), ind = NA, var = NA)
}
\arguments{
\item{model}{an sfc object.}

\item{ind}{index of equation to be removed.}
}
\value{
an sfc object.
}
\description{
Remove an equation from an sfc model.
}
\author{
Antoine Godin
}

