% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.rmVar}
\alias{sfc.rmVar}
\title{Remove a variable.}
\usage{
sfc.rmVar(model = stop("Need a model"), ind = stop("Need an index!"))
}
\arguments{
\item{model}{an sfc object.}

\item{ind}{index of the variable to be edited.}
}
\value{
an sfc object.
}
\description{
Remove a variable from an sfc object.
}
\author{
Antoine Godin
}

