% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sfc.setYear}
\alias{sfc.setYear}
\title{Set the years in a sfc object.}
\usage{
sfc.setYear(model = stop("Need a model"),
  init = stop("Need an initial year"), end = stop("Need and ending year"))
}
\arguments{
\item{model}{an sfc object.}

\item{init}{intial date.}

\item{end}{end date.}
}
\value{
an sfc object.
}
\description{
Set the years for which the sfc object is to examined over.
}
\author{
Antoine Godin
}

