% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate.sfc}
\alias{simulate.sfc}
\title{Simulating SFC models}
\usage{
\method{simulate}{sfc}(model, tolValue = 1e-10, maxIter = 10000)
}
\arguments{
\item{model}{contains the sfc model to be simulated}

\item{tolValue}{minimum variation between two iteration of the Gauss-Seidel
algorithm.}

\item{maxIter}{maximum number of iteration for the Gauss-Seidel algorithm}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Simulate using the Gaiss-Seodel algorithm an sfc model.
}
\details{
Simulates, via the Gauss-Seidel algorithm, an SFC model for the baseline and
all scenarios that where added. See sfc.addScenario for more on scenarios
and sfc.GaussSeidel for more on the Gauss-Seidel algorithm.
ore details than the description above ~~
}
\author{
Antoine Godin
}

