% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.sfc}
\alias{summary.sfc}
\title{Create a summary of an sfc object}
\usage{
\method{summary}{sfc}(object, ...)
}
\arguments{
\item{object}{The sfc object to be summarised}

\item{...}{optional parameters to be passed to function}
}
\value{
A summary of the sfc object
}
\description{
Create a summary of an sfc object
}
\author{
Antoine Godin
}

