% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test.node.removal}
\alias{test.node.removal}
\title{Test to see if removing certain nodes will result in a DAG}
\usage{
test.node.removal(adjacency, nodes_to_remove, root_node = 1, PLOT = FALSE)
}
\arguments{
\item{adjacency}{A square adjacency matrix.}

\item{nodes_to_remove}{A vector of nodes to remove}

\item{root_node}{The node that forms the root of the plot.}

\item{PLOT}{If plot is \code{TRUE}, create a plot of the graph with nodes that are to be removed
highlighted in light pink and nodes outside of cycles higlighted in green.}
}
\value{
Returns \code{TRUE} if removing nodes results in a DAG, othereise \code{FALSE} is
returned.
}
\description{
Test to see if removing certain nodes will result in a DAG
}
\examples{
data( models )
test.node.removal( sim$matrix, 1 ) # test to see if removing the first node will create a DAG
}
\author{
David O'Sullivan
}

